Get WuGe information of Chinese names

`wuge()` gets the WuGe information of Chinese names.

Usage

wuge(xing, ming, to_trad = TRUE)

Arguments

xing

A character vector of Chinese characters indicating the last names.

ming

A character vector of Chinese characters indicating the first names.

to_trad

If `TRUE`, input characters are converted to Traditional Chinese using [sim2trad] before calculation. If [sim2trad] does not give correct results, you may want to set `to_trad` to FALSE and directly specify Tradition Chinese characters as input. Default: `TRUE`

Value

An `WuGe` object.

Examples

wuge(c("张", "李"), c("三", "四五"))
#> ══ 姓名 ════════════════════════════════════════════════════════════════ 张三 ══
#> • 姓名：   张  三
#> • 繁体：   張  三
#> • 拼音：zhānɡ sān
#> • 笔画：   11   3
#> 
#> ── 五格数理排盘 ──────────────────────────────────────────────────────── 张三 ──
#> ┌─────────┐
#> │ 46.5 分 │
#> └─────────┘
#> • 天格：12 -> 木 ->   凶
#> • 人格：14 -> 火 ->   凶
#> • 地格： 4 -> 火 ->   凶
#> • 外格： 2 -> 木 -> 大凶
#> • 总格：14 -> 火 ->   凶
#> 
#> ── 天格 12 的分析 ────────────────────────────────────────────────────── 张三 ──
#> ┌───────┐
#> │ 55 分 │
#> └───────┘
#> • 关键词：挖井无泉(掘井无泉)
#> • 简述：无理之数，发展薄弱，虽生不足，难酬志向。
#> • 数理暗示：
#>   • 意志脆弱，家庭寂寞数
#>   • 凶数运（表示逆境、沉浮、薄弱、困难、多灾等）。
#>   • 孤独运（妻不利夫或夫不利妻）。
#> • 基础运：凶星、破危、劫财、时禄（一段好运）。
#> • 家庭运：亲情如秋水，骨肉似寒炭，施恩招怨恨，宜提高气节。
#> • 健康运：枯木待春，小心神经衰弱，肝、胃病之疾、多伤（害），皮肤病，三才善良者，可望安全。
#> • 事业运：薄弱无力、孤立无摇、外祥内苦、谋事难成。
#> • 财富运：君臣之像，心比天高，自不量力，劳累奔波，有财易大起大落。
#> • 详细解释：
#>   (1)有无理伸张之意，不顾脆弱之力，企图做无能力做的事，多失败。退事易生不足之心，家庭无缘、孤独、遇难、逆境、病弱、不如意、困难等，又可有其他运配合而导致意外的失败，甚至有不能完寿的悲运。
#>   (2)做事经不起考验，容易患得患失，遇事则易半途而废，错失良机而后悔莫及，凡事不堪打击，一败涂地，自怜自艾。常图非份之心，招致失败，以致陷入苦境。家族缘甚薄。个性温文儒雅，受朋友欢迎及长辈提拔，但意志薄弱，身体差，无法居领导位置，宜过安稳固定之生活。
#> 
#> ── 人格 14 的分析 ────────────────────────────────────────────────────── 张三 ──
#> ┌───────┐
#> │ 45 分 │
#> └───────┘
#> • 关键词：破兆
#> • 简述：家庭缘薄，孤独遭难，谋事不达，悲惨不测。
#> • 数理暗示：
#>   • 沦落天崖的失意、烦闷数
#>   • 凶数运（表示逆境、沉浮、薄弱、困难、多灾等）。
#>   • 艺能运（富有艺术天才，对审美、艺术、演艺、体育有通达之能）。
#>   • 孤独运（妻不利夫或夫不利妻）。
#> • 基础运：暗禄、美貌、艺术、流浪、红舞、劫财。事业半途而废（不愿做完或开除、失业、下岗等））。
#> • 家庭运：骨肉疏远，离祖迁居，自家兄弟全无缘分，外处事有缘分。
#> • 健康运：男女均属好貌或身材（才）。皮肤病、感冒、风湿病，三才配置善良，可保健康。
#> • 事业运：吃得苦难、必有后福、是成是败、惟靠坚毅。
#> • 财富运：虽有意外收获，易被他人牵连祸害，先甜后苦，大起大落。
#> • 详细解释：
#>   (1)多破兆，家属缘薄。丧亲、丧子（或无子），兄弟姐妹分离孤独，不如意、烦闷、危难、遭恶、浮沉不定，为了慷慨，施惠招怨，劳而无功，凄惨孤独，其他运者不多，有伤天寿。
#>   (2)常为经济拮据而困扰，因本位主义太重，不听劝告，一意孤行，为人好虚荣，重外表，不充实内涵，以致诸事不能如意，一生暗淡，家族缘薄，孤独无依，亲友缘薄不得助，孤僻冷漠，为人高傲，所以生活方面很苦闷。
#> 
#> ── 地格 4 的分析 ─────────────────────────────────────────────────────── 张三 ──
#> ┌───────┐
#> │ 50 分 │
#> └───────┘
#> • 关键词：凶变(四象之数)
#> • 简述：四象之数，待于生发，万事慎重，不具营谋。
#> • 数理暗示：
#>   • 身遭凶变的万事休止数
#>   • 凶数运（表示逆境、沉浮、薄弱、困难、多灾等）。
#>   • 孤独运（妻不利夫或夫不利妻）。
#> • 基础运：美貌（良心好）、香艳（公认的、有名气的）、破家、灾厄、劫财、败家业。
#> • 家庭运：六亲缘薄、兄弟如同画饼。热先相助的少、应看开点。
#> • 健康运：衰弱、外伤、皮肤病、夭折、病死、发狂病（心里不平衡所致）。
#> • 事业运：坎坷前途、苦难折磨、非有毅力、望难成功。
#> • 财富运：天时人和只欠地利，意外收获，一时发展，易大起大落。
#> • 详细解释：
#>   (1)破坏性的凶变数，不足不全的灭亡之数，进退不自由、独立无能力，大多辛苦困难，或者逢其他凶运的配合而致发狂病死、夭折，或者放荡、破灭、逆难、终成废人，但孝子、节妇、怪杰，也有出自此数者。
#>   (2)此数为不伦不类、无能生存之数，为人大多消极，性阴沉顽固，难成大事。一生精神多苦闷，力不从心，虽有才华也很努力开创，但总是脱离现实，很难被认同；身体尚佳，财运无缺，但对人际关系的处理上，较不圆融，易受亲友误解。不宜从事投机冒险事业，以免遭受失败。
#> 
#> ── 外格 2 的分析 ─────────────────────────────────────────────────────── 张三 ──
#> ┌───────┐
#> │ 45 分 │
#> └───────┘
#> • 关键词：一身孤节(两仪之数)
#> • 简述：两仪之数，混沌未开，进退保守，志望难达。
#> • 数理暗示：
#>   • 未定的分离破灭数
#>   • 凶数运（表示逆境、沉浮、薄弱、困难、多灾等）。
#> • 基础运：劫财、破灭、灾难危险、红颜，变迁、美貌
#> • 家庭运：亲情疏远，夫妻应该互相理解，则免离别之苦。
#> • 健康运：凶弱、短命定有的，易患皮肤病，外伤，智折。
#> • 事业运：根基不固、遥遥欲坠、一盛一衰、劳而无获。
#> • 财富运：天时地利只欠人和，一生辛苦、劳累奔波、大起大落，难得收获。
#> • 详细解释：
#>   (1)混纯未定的数，为最大凶恶的暗示，无独立之气力，进退无自由，内外波澜、困苦不安，无雄心，易动摇，病患、遭难、甚至残废。如三才善良者不会夭折，其人辛苦一生（身），志望难达，破灭无常。（逆境、薄弱、病难、沉浮、短命、凶杀、非命等）
#>   (2)此数为万事万物生长之数，但有形无气，有柔顺、无力之意，为人性柔顺，重利，被动，欠缺创新的思考能力，个性叛逆，遇事坚持己见，但无独立气魄。
#> 
#> ── 总格 14 的分析 ────────────────────────────────────────────────────── 张三 ──
#> ┌───────┐
#> │ 45 分 │
#> └───────┘
#> • 关键词：破兆
#> • 简述：家庭缘薄，孤独遭难，谋事不达，悲惨不测。
#> • 数理暗示：
#>   • 沦落天崖的失意、烦闷数
#>   • 凶数运（表示逆境、沉浮、薄弱、困难、多灾等）。
#>   • 艺能运（富有艺术天才，对审美、艺术、演艺、体育有通达之能）。
#>   • 孤独运（妻不利夫或夫不利妻）。
#> • 基础运：暗禄、美貌、艺术、流浪、红舞、劫财。事业半途而废（不愿做完或开除、失业、下岗等））。
#> • 家庭运：骨肉疏远，离祖迁居，自家兄弟全无缘分，外处事有缘分。
#> • 健康运：男女均属好貌或身材（才）。皮肤病、感冒、风湿病，三才配置善良，可保健康。
#> • 事业运：吃得苦难、必有后福、是成是败、惟靠坚毅。
#> • 财富运：虽有意外收获，易被他人牵连祸害，先甜后苦，大起大落。
#> • 详细解释：
#>   (1)多破兆，家属缘薄。丧亲、丧子（或无子），兄弟姐妹分离孤独，不如意、烦闷、危难、遭恶、浮沉不定，为了慷慨，施惠招怨，劳而无功，凄惨孤独，其他运者不多，有伤天寿。
#>   (2)常为经济拮据而困扰，因本位主义太重，不听劝告，一意孤行，为人好虚荣，重外表，不充实内涵，以致诸事不能如意，一生暗淡，家族缘薄，孤独无依，亲友缘薄不得助，孤僻冷漠，为人高傲，所以生活方面很苦闷。
#> 
#> ── 三才剖析 ──────────────────────────────────────────────────────────── 张三 ──
#> ┌ 木火火 ──────────────┐
#> │ 吉凶：<平>；得分：75 │
#> └──────────────────────┘
#> ── 简述 ──────────────────────────────────────────────────────────────── 张三 ──
#>     得顺调成功发展，但有缺乏耐久力的缺点，或依靠性太强，招失败，有陷入失意之兆。
#>     起初难免有孤革奋斗之感，但斗志激昂，终排除困难，达到顺调，成功发达，尤能得上位之惠肋而更伸张发展。但似乎乏其持久，耐久力，以致虽是成败交加频见，总之大体还算安然，但须防火灾或烫伤之灾。人地格皆23数：甚好淫。
#> 
#> ── 详细解释 ──────────────────────────────────────────────────────────── 张三 ──
#> • 总论：虽有成功运，但不甘寂寞，忍耐力不够，容易一意孤行，陷入孤军奋斗，如能修身养性，定能成功顺调。名利双利，若天运五行属水，青年时代事业难成又逢灾，老运亦欠佳。
#> • 性格：外表乐观而豪爽，交际应酬有一套，人缘又好，但个性喜怒无常，感情易热易冷，个性急燥，容易意气用事。
#> • 意志：意志尚称坚定，可依自己的理想和努力，达到成功成名，但不耐寂寞独处生活。
#> • 事业：做事冲动而积极，很容易功成名就，一展自己的抱负。
#> • 家庭：夫妻美满，子女孝顺，但应节制应酬才能保持长久。
#> • 婚姻：男娶热情直率之妻，应减少应酬风流之事；女嫁温厚谦恭之夫，要减少让丈夫多疑之行为。
#> • 子女：男孩多于女孩，每个都健康活泼，稍嫌调皮，长大后都能获得相当成就。
#> • 社交：人缘很好，朋友很多，易与异性接触，避免桃色事件发生。
#> • 精神：外表乐观，内在则脾气暴，常自找精神上的负担。
#> • 财运：物质不缺，财源广进，富贵之格。
#> • 健康：注意血压过高或心脏病。
#> • 老运：可享受晚年，但恐缺天份之乐。
#> 
#> ── 基础运 ────────────────────────────────────────────────────────────── 张三 ──
#> ┌────────────┐
#> │ 吉凶：<吉> │
#> └────────────┘
#> • 五格：人格【火】地格【火】
#> • 描述：可得一时的吉运，但基础薄弱，缺乏耐久力，但天格为1或2时则为吉利。
#> 
#> ── 成功运 ────────────────────────────────────────────────────────────── 张三 ──
#> ┌────────────┐
#> │ 吉凶：<吉> │
#> └────────────┘
#> • 五格：天格【木】人格【火】
#> • 描述：受上级长辈的照顾、关心，能顺利成功发达。
#> 
#> ── 社交运 ────────────────────────────────────────────────────────────── 张三 ──
#> ┌────────────┐
#> │ 吉凶：<吉> │
#> └────────────┘
#> • 五格：人格【4】外格【2】
#> • 描述：热情好礼乐善好施，表面乐观，内实劳苦，而24数理在人格者可白手起家，24数以外易有胃腹及外伤。
#> 
#> ══ 姓名 ══════════════════════════════════════════════════════════════ 李四五 ══
#> • 姓名：李 四 五
#> • 繁体：李 四 五
#> • 拼音：lǐ sì wǔ
#> • 笔画： 7  4  5
#> 
#> ── 五格数理排盘 ────────────────────────────────────────────────────── 李四五 ──
#> ┌──────────┐
#> │ 86.85 分 │
#> └──────────┘
#> • 天格： 8 -> 金 ->   吉
#> • 人格：11 -> 木 -> 大吉
#> • 地格： 9 -> 水 ->   凶
#> • 外格： 6 -> 土 -> 半吉
#> • 总格：16 -> 土 -> 大吉
#> 
#> ── 天格 8 的分析 ───────────────────────────────────────────────────── 李四五 ──
#> ┌───────┐
#> │ 85 分 │
#> └───────┘
#> • 关键词：坚刚(八卦之数)
#> • 简述：八卦之数，乾坎艮震，巽离坤兑，无穷无尽。
#> • 数理暗示：
#>   • 志刚意健的勤勉发展数
#>   • 吉祥运（表示健全、幸福、名誉等）。
#> • 基础运：艺能、美术、学者、天（官）、（官）禄、师表、畜产（喜养殖）。
#> • 家庭运：兴家成为发达贤人，家境先能后易。
#> • 健康运：高山立松完健百在，可望长寿。
#> • 事业运：努力发达、贯彻志望、不忘进退、可望成功。
#> • 财富运：天时地利人和，应注意修和，努力进取，财富必得。
#> • 详细解释：
#>   (1)意志如钢铁，富于进取的气质，排除万难。贯彻目的，名、实两得、忍耐而克己，逐成大功，但其他运配合不善者，可能有遭难的厄运。慎防官司、牢狱之灾。
#>   (2)具有坚忍不拔的恒心和毅力，能贯彻意志成就愿望。但有时坚持己见，却容易迷失方向。具超强意志力和耐力，看来精力旺盛不服输，总是能突破重重难关，获得成功，拥有此数，大都少年早发；是一个宁为鸡头不为凤尾之人，希望自己施展抱负。一生容易成功，但也容易骄傲自满、自大，而引人反感而招来烦恼。
#> 
#> ── 人格 11 的分析 ──────────────────────────────────────────────────── 李四五 ──
#> ┌───────┐
#> │ 98 分 │
#> └───────┘
#> • 关键词：旱苗逢雨
#> • 简述：万物更新，调顺发达，恢弘泽世，繁荣富贵。
#> • 数理暗示：
#>   • 挽回家运的成育数
#>   • 吉祥运（表示健全、幸福、名誉等）。
#>   • 温和运（性情平和、能得上下信望）。
#> • 基础运：财星、天使、暗禄、文昌、技艺、田宅。
#> • 家庭运：养蜂结蜜，事事和顺，处处温和。
#> • 健康运：河川永在，可望健康长寿。
#> • 事业运：草木逢春、枝叶沾露、稳健着实、必得人望。
#> • 财富运：天时地利人和，有贵相助，名利双收，步步高升，荣华富贵。
#> • 详细解释：
#>   (1)享有天赋幸运，万事顺利，有得富贵荣誉，再兴家业的暗示，得挽回家运、平静和顺的最大吉数。
#>   (2)看来是吉星高照，逢凶化吉之数，一生能平顺通达，福财禄不缺，很早就能享名望，一生敢做敢当，脚踏实地，深受同业或朋友敬佩。此数也是草木逢春、处处生机、天资聪颖之数，为人平稳踏实，得名望，当可大展鸿图。但主观意念稍重，易陷独断独行，反招人怨。
#> 
#> ── 地格 9 的分析 ───────────────────────────────────────────────────── 李四五 ──
#> ┌───────┐
#> │ 50 分 │
#> └───────┘
#> • 关键词：破舟进海(大成之数)
#> • 简述：大成之数，蕴涵凶险，或成或败，难以把握。
#> • 数理暗示：
#>   • 吉尽凶始，穷乏之星
#>   • 凶数运（表示逆境、沉浮、薄弱、困难、多灾等）。
#> • 基础运：(官)禄、图印、太极、富翁、突（然）破（家或产）
#> • 家庭运：亲情不和睦，说长道短的人多，努力坚持贤明者平安。
#> • 健康运：女性成风流或与夫别离，病弱。弱女者都应该晚婚，多积德可免短命。
#> • 事业运：虽抱奇才、有才无命、独营无力、财力难忘。
#> • 财富运：聪明智慧、财路很多，贪功好进，大起大落。如能行善，财富很多。
#> • 详细解释：
#>   (1)利去皆空，陷入穷途，逆运、短命、悲痛或幼小离亲而困苦、或者病弱、不遇、遭难废疾、贫困、灾难、孤独。甚至有刑罚、有不测之凶祸。人格有此数者为大凶或一生得免灾害，也难免配偶丧失，有缺子之叹，实为人生最大恶运，但在，例外的怪杰、富豪也有出此数者。
#>   (2)一生虽理想高，但容易失败，有时有志难伸，知音难寻，容易超越现实不易融于社会。意志不很坚定，行动举止比较轻浮，是为孤苦无依之数，一生恋爱运、亲友运差，较难得到身边之人的帮助，且常陷于是非纠纷之中。倘能在失败挫折中潜心学习，或能在艺术、宗教、五术哲学领域中，渐渐得到成功发展。
#> 
#> ── 外格 6 的分析 ───────────────────────────────────────────────────── 李四五 ──
#> ┌───────┐
#> │ 80 分 │
#> └───────┘
#> • 关键词：安稳(六爻之数)
#> • 简述：六爻之数，发展变化，天赋美德，吉祥安泰。
#> • 数理暗示：
#>   • 吉庆的吉人天相数
#>   • 次吉祥运（表示多少有些障碍，但能获得吉运）。
#>   • 男性双妻运（异性缘旺盛）。
#>   • 温和运（性情平和、能得上下信望）。
#> • 基础运：豪杰、（官）禄、财富、（将）星、学者、
#> • 家庭运：逢凶化吉、有一次大风险（危险）、可得长寿。
#> • 健康运：安宁自在，家庭和睦，可望健康。
#> • 事业运：万宝集门、天降幸运、立志奋发、得成大功。
#> • 财富运：应讲信修和、积善行德、不至于人财两空，必得人旺，吉庆一生。
#> • 详细解释：
#>   (1)天德地祥，福庆甚丰，家势盛大，万宝朝宗之运。然而满极则损，盈极则亏。若其他数理配合不周者，恐相如流水而下，成为所谓乐极生悲之相，此数理俱有天赋之美，安稳吉庆终身。
#>   (2)为人豪爽，花钱较无节制，此数之人聪慧，兼具有美德之性格，若能平稳持重，可享和乐长寿，不然则容易乐极生悲，反为不利，但不至于会困顿潦倒。一生平顺，容易享有祖上余荫，具长者风范，愿提携后进，深得部属欢心喜爱；较爱面子，用钱较浪费，但在社会上，容易享有高声望。看来桃花多，婚后常有应酬，宜节制。
#> 
#> ── 总格 16 的分析 ──────────────────────────────────────────────────── 李四五 ──
#> ┌───────┐
#> │ 98 分 │
#> └───────┘
#> • 关键词：厚重
#> • 简述：厚德载物，安富尊荣，财官双美，功成名就。
#> • 数理暗示：
#>   • 兴家得助的贵人数
#>   • 首领运（智慧仁勇、立上位、能领导众人）。
#>   • 富贵运（多钱财、富贵、白手可获巨财）。
#>   • 吉祥运（表示健全、幸福、名誉等）。
#>   • 男性双妻运（异性缘旺盛）。
#>   • 温和运（性情平和、能得上下信望）。
#> • 基础运：天（官）、贵（人）命、豪杰、进田、学士、豪侠、置产。
#> • 家庭运：春月花开，可望家庭圆满，男子有贤妻，女子不宜早婚。
#> • 健康运：花草逢春，贵弱于色情，戒慎者可望健康。
#> • 事业运：能获众望、成就大业、名利双收、盟主四方。
#> • 财富运：财涌自然来，天赋亨通，富裕荣华，名利双收。
#> • 详细解释：
#>   (1)反凶化育之相，成首领，占上峰，雅量厚重，豪侠足智多谋，善于协调，集名望，能服从，为大事大业可成，富贵发达的好暗示。
#>   (2)为人较爱面子，一生奋发，有果断力，能任劳任怨不辞劳苦，容易成就财富与功名，而受众人拥戴、推崇，又能获长上器重，而得到名望，成就一番事业，是富贵吉祥之数。与人谦恭有礼也乐于助人，人缘好，乐观幽默，深得异性喜爱；精力充沛，耐性也够。
#> 
#> ── 三才剖析 ────────────────────────────────────────────────────────── 李四五 ──
#> ┌ 金木水 ──────────────┐
#> │ 吉凶：<凶>；得分：35 │
#> └──────────────────────┘
#> ── 简述 ────────────────────────────────────────────────────────────── 李四五 ──
#>     成功运被压抑，难以伸张，易失败流转。
#>     颇有一时之成就而博得名利，但因成功运被压制，不能有所伸张发展，恐易再陷落流亡、失败、逆境、流浪或遭电击所伤、武器杀伤，甚至变为短寿或染有难治之顽疾，尤其防车祸。
#> 
#> ── 详细解释 ────────────────────────────────────────────────────────── 李四五 ──
#> • 总论：早年有不祥之兆，凭努力及耐力，中年可成功发展，但事业有盛衰起落之灵意，处事应多谨慎注意。天运五行属火之人，中年后财运甚丰。
#> • 性格：行事稳重可靠，注意实际与工作，肯为他人牺牲，是讲义气的类型，但有时对自己满怀信心，容易陷入一意孤行，应改进。
#> • 意志：意志坚定，靠自己的毅力突破万难，耐心亦佳，能忍受艰苦，是创业的人才。
#> • 事业：中年前事业阻挠很多，如能经得起考验，中年后有成就事业之象，但须忍耐导重行事。
#> • 家庭：父母无荫，亲友难助，夫妻缘薄，子女亦难靠。
#> • 婚姻：男娶克勤克俭之妻，婚后家内不安；女嫁好胜寡和之夫，婚后多争执。
#> • 子女：女孩多于男孩，性格较孤独，身体状况不佳。
#> • 社交：人缘及社交能力还不错，肯为人操劳，但主观较强，容易被人误会，须改进。
#> • 精神：表面似乎乐观，但理想太高，达成不易，精神烦闷不安。
#> • 财运：财运不稳定，时贫时富，财务问题应慎重处置。
#> • 健康：易患神经衰弱、肠胃、头痛、肾脏机能、妇人病等症。
#> • 老运：一生劳碌，晚景较安定，但家内欠温暖，精神烦闷不安。
#> 
#> ── 基础运 ──────────────────────────────────────────────────────────── 李四五 ──
#> ┌────────────┐
#> │ 吉凶：<凶> │
#> └────────────┘
#> • 五格：人格【木】地格【水】
#> • 描述：一时虽能顺利发展，但不时可能有衰亡，病弱的症兆。
#> 
#> ── 成功运 ──────────────────────────────────────────────────────────── 李四五 ──
#> ┌────────────┐
#> │ 吉凶：<凶> │
#> └────────────┘
#> • 五格：天格【金】人格【木】
#> • 描述：命运被庄抑，导致难伸不满的结果，胸部被克，患神经衰弱，呼吸器官等难治之症。
#> 
#> ── 社交运 ──────────────────────────────────────────────────────────── 李四五 ──
#> ┌──────────────┐
#> │ 吉凶：<半吉> │
#> └──────────────┘
#> • 五格：人格【1】外格【6】
#> • 描述：好奇心强，但不求实干，如不自轻心，则可望成功，言行有言的人，不会有大过失，多患近视、皮肤病，不宜与人合作。